Split modlist generator and uploader
