+++ .github/workflows/modlist.yml
@@ -15,4 +15,5 @@ jobs:
         run: cd automation/node && npm install
 
 #      - run: echo '${{ toJSON(github) }}'
-      - run: node automation/node/new-modlist.js ${{ github.token }} ${{ github.event.repository.owner.name }} ${{ github.event.repository.name }} ${{ github.ref_name }}
+      - run: node automation/node/modlist-create.js
+      - run: node automation/node/modlist-commit.js ${{ github.token }} ${{ github.event.repository.owner.name }} ${{ github.event.repository.name }} ${{ github.ref_name }}
